requirements.txt
Django==1.8.1
django-model-utils==2.2
djangorestframework==3.1.2
